Move code from `MediaFormatter` to `LineBlockFormatter`
